requirements.txt
customtkinter==5.2.2
tkinterdnd2>=0.3.0
openpyxl==3.1.2
reportlab==4.2.0
Pillow==10.3.0
PyMuPDF>=1.24.0
pywin32==306
pyinstaller==6.6.0
